# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: PesenHub API
  version: 1.0.0
  description: Shared HTTP primitives for Phase 1A. Domain paths are added by their owning issues.
servers:
  - url: /api/v1
paths:
...
  /admin/categories:
    post:
      summary: Create a category
      operationId: createCategory
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CategoryMutation'}}}
      responses:
        '201': {description: Category created, content: {application/json: {schema: {$ref: '#/components/schemas/CatalogCategory'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/ValidationError'}
...
  /admin/categories/{id}:
    patch:
      summary: Update category fields with optimistic version
      operationId: updateCategory
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CategoryMutation'}}}
      responses:
        '200': {description: Category updated, content: {application/json: {schema: {$ref: '#/components/schemas/CatalogCategory'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /admin/menus:
    post:
      summary: Create a menu with modifier groups and options atomically
      operationId: createMenu
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/MenuMutation'}}}
      responses:
        '201': {description: Menu created, content: {application/json: {schema: {$ref: '#/components/schemas/CatalogMenu'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /admin/menus/{id}:
    patch:
      summary: Update menu, price, ordering, and modifier tree with optimistic version
      operationId: updateMenu
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/MenuMutation'}}}
      responses:
        '200': {description: Menu updated, content: {application/json: {schema: {$ref: '#/components/schemas/CatalogMenu'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
...
  /orders:
    get:
      summary: List and filter orders with cursor pagination
      operationId: listOrders
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/Sort'
        - name: status
          in: query
          description: Filter by status (single or comma-separated)
          schema: {type: string}
        - name: source
          in: query
          description: Filter by order source (single or comma-separated)
          schema: {type: string}
        - name: created_from
          in: query
          description: Filter by creation time start (RFC3339)
          schema: {type: string, format: date-time}
        - name: created_to
          in: query
          description: Filter by creation time end (RFC3339)
          schema: {type: string, format: date-time}
      responses:
        '200': {description: Filtered orders with cursor pagination, content: {application/json: {schema: {$ref: '#/components/schemas/OrderCollection'}}}}
        '400': {$ref: '#/components/responses/InvalidQuery'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/ValidationError'}
    post:
      summary: Create or replay a cashier manual order atomically
      operationId: createManualOrder
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ManualOrderCreate'}}}
      responses:
        '200': {description: Identical idempotent replay, content: {application/json: {schema: {$ref: '#/components/schemas/Order'}}}}
        '201': {description: Pending cashier order created, content: {application/json: {schema: {$ref: '#/components/schemas/Order'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
...
components:
  securitySchemes:
    gowaWebhookHMAC:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: Lowercase hex HMAC-SHA256 of the unmodified raw request body, prefixed with sha256=.
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      schema:
        type: string
        pattern: '^[A-Za-z0-9._-]{1,128}$'
    PageSize:
      name: page[size]
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
    PageCursor:
      name: page[cursor]
      in: query
      schema:
        type: string
        description: Opaque base64url cursor returned by the server.
    Sort:
      name: sort
      in: query
      schema:
        type: string
        enum: [created_at, -created_at]
        default: created_at
  headers:
    RequestID:
      description: Effective correlation identifier.
      schema:
        type: string
  responses:
    ValidationError:
      description: Semantic input validation failed.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    Conflict:
      description: Idempotency, phone identity, or version conflict.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    Forbidden:
      description: Authenticated principal cannot access this customer.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    InvalidQuery:
      description: Query parameter validation failed.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    ModifierSelection:
      type: object
      additionalProperties: false
      required: [group_id, option_ids]
      properties:
        group_id: {type: string, format: uuid}
        option_ids: {type: array, items: {type: string, format: uuid}}
    ManualOrderItemCreate:
      type: object
      additionalProperties: false
      required: [menu_id, quantity]
      properties:
        menu_id: {type: string, format: uuid}
        quantity: {type: integer, minimum: 1, maximum: 100}
        notes: {type: string}
        modifier_groups: {type: array, items: {$ref: '#/components/schemas/ModifierSelection'}}
    ManualOrderCreate:
      type: object
      additionalProperties: false
      required: [client_order_id, customer_name, items]
      properties:
        client_order_id: {type: string, format: uuid}
        customer_id: {type: string, format: uuid}
        customer_name: {type: string, minLength: 1, maxLength: 120}
        customer_phone: {type: string}
        notes: {type: string}
        items: {type: array, minItems: 1, maxItems: 100, items: {$ref: '#/components/schemas/ManualOrderItemCreate'}}
    OrderItem:
      type: object
      required: [id, menu_id, name, sku, unit_price_amount, quantity, line_total_amount]
      properties:
        id: {type: string, format: uuid}
        menu_id: {type: string, format: uuid}
        name: {type: string}
        sku: {type: string}
        unit_price_amount: {type: integer, format: int64, minimum: 0}
        quantity: {type: integer, minimum: 1}
        line_total_amount: {type: integer, format: int64, minimum: 0}
    Order:
      type: object
      required: [id, order_number, client_order_id, source, status, total_amount, version, created_at, items]
      properties:
        id: {type: string, format: uuid}
        order_number: {type: string}
        client_order_id: {type: string, format: uuid}
        source: {const: CASHIER_MANUAL}
        status: {const: PENDING}
        total_amount: {type: integer, format: int64, minimum: 0}
        version: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
        items: {type: array, items: {$ref: '#/components/schemas/OrderItem'}}
...
    ModifierSnapshot:
      type: object
      required: [id, name, price_delta_amount]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        price_delta_amount: {type: integer, format: int64}
    OrderItemDetail:
      type: object
      required: [id, menu_id, name, sku, unit_price_amount, quantity, line_total_amount]
      properties:
        id: {type: string, format: uuid}
        menu_id: {type: string, format: uuid}
        name: {type: string}
        sku: {type: string}
        category_name: {type: string}
        quantity: {type: integer, minimum: 1}
        unit_price_amount: {type: integer, format: int64, minimum: 0}
        line_total_amount: {type: integer, format: int64, minimum: 0}
        notes: {type: string}
        modifiers:
          type: array
          items: {$ref: '#/components/schemas/ModifierSnapshot'}
    OrderStatusHistoryEntry:
      type: object
      required: [to_status, order_version, actor_type, created_at]
      properties:
        from_status: {type: string}
        to_status: {type: string}
        order_version: {type: integer, minimum: 1}
        actor_type: {type: string}
        actor_id: {type: string}
        reason_code: {type: string}
        created_at: {type: string, format: date-time}
    OrderDetail:
      type: object
      required: [id, order_number, source, status, customer_name, total_amount, version, created_at, updated_at, items]
      properties:
        id: {type: string, format: uuid}
        order_number: {type: string}
        client_order_id: {type: string, format: uuid}
        customer_id: {type: string, format: uuid}
        source: {type: string, enum: [CASHIER_MANUAL, CUSTOMER_WEB, WHATSAPP]}
        status: {type: string, enum: [PENDING, ACCEPTED, PREPARING, READY_FOR_PICKUP, COMPLETED, REJECTED, CANCELLED]}
        customer_name: {type: string}
        customer_phone: {type: string, nullable: true}
        notes: {type: string}
        total_amount: {type: integer, format: int64, minimum: 0}
        version: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        items:
          type: array
          items: {$ref: '#/components/schemas/OrderItemDetail'}
        history:
          type: array
          items: {$ref: '#/components/schemas/OrderStatusHistoryEntry'}
    OrderCollection:
      type: object
      additionalProperties: false
      required: [data, page]
      properties:
        data:
          type: array
          items: {$ref: '#/components/schemas/OrderDetail'}
        page:
          $ref: '#/components/schemas/PageMeta'
...
    ModifierOption:
      type: object
      required: [id, code, name, price_delta_amount, is_available, sort_order]
      properties:
        id: {type: string, format: uuid}
        code: {type: string}
        name: {type: string}
        price_delta_amount: {type: integer, format: int64}
        is_available: {type: boolean}
        sort_order: {type: integer, minimum: 0}
    ModifierGroup:
      type: object
      required: [id, code, name, min_select, max_select, is_active, sort_order, options]
      properties:
        id: {type: string, format: uuid}
        code: {type: string}
        name: {type: string}
        min_select: {type: integer, minimum: 0}
        max_select: {type: integer, minimum: 1}
        is_active: {type: boolean}
        sort_order: {type: integer, minimum: 0}
        options: {type: array, items: {$ref: '#/components/schemas/ModifierOption'}}
    CatalogMenu:
      type: object
      required: [id, category_id, sku, name, price_amount, is_available, version, sort_order, modifier_groups]
      properties:
        id: {type: string, format: uuid}
        category_id: {type: string, format: uuid}
        sku: {type: string}
        name: {type: string}
        price_amount: {type: integer, format: int64, minimum: 0}
        is_available: {type: boolean}
        version: {type: integer, minimum: 1}
        sort_order: {type: integer, minimum: 0}
        modifier_groups: {type: array, items: {$ref: '#/components/schemas/ModifierGroup'}}
    CatalogCategory:
      type: object
      required: [id, name, sort_order, is_active, version, menus]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        sort_order: {type: integer, minimum: 0}
        is_active: {type: boolean}
        version: {type: integer, minimum: 1}
        menus: {type: array, items: {$ref: '#/components/schemas/CatalogMenu'}}
    CategoryMutation:
      type: object
      additionalProperties: false
      required: [name, sort_order, is_active, version]
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        sort_order: {type: integer, minimum: 0}
        is_active: {type: boolean}
        version: {type: integer, minimum: 1}
    ModifierOptionMutation:
      type: object
      additionalProperties: false
      required: [code, name, price_delta_amount, sort_order]
      properties:
        code: {type: string, minLength: 1, maxLength: 64}
        name: {type: string, minLength: 1, maxLength: 120}
        price_delta_amount: {type: integer, format: int64}
        sort_order: {type: integer, minimum: 0}
    ModifierGroupMutation:
      type: object
      additionalProperties: false
      required: [code, name, min_select, max_select, sort_order, options]
      properties:
        code: {type: string, minLength: 1, maxLength: 64}
        name: {type: string, minLength: 1, maxLength: 120}
        min_select: {type: integer, minimum: 0}
        max_select: {type: integer, minimum: 1}
        sort_order: {type: integer, minimum: 0}
        options: {type: array, items: {$ref: '#/components/schemas/ModifierOptionMutation'}}
    MenuMutation:
      type: object
      additionalProperties: false
      required: [category_id, sku, name, price_amount, sort_order, version, modifier_groups]
      properties:
        category_id: {type: string, format: uuid}
        sku: {type: string, minLength: 1, maxLength: 64}
        name: {type: string, minLength: 1, maxLength: 160}
        description: {type: string}
        price_amount: {type: integer, format: int64, minimum: 0}
        sort_order: {type: integer, minimum: 0}
        version: {type: integer, minimum: 1}
        modifier_groups: {type: array, items: {$ref: '#/components/schemas/ModifierGroupMutation'}}
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        request_id:
          type: string
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
    PageMeta:
      type: object
      additionalProperties: false
      required: [size, next_cursor]
      properties:
        size:
          type: integer
          minimum: 1
          maximum: 100
        next_cursor:
          type: [string, 'null']
...